+++ robot_descriptions/yam_description.py
@@ -1,0 +1,21 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+#
+# SPDX-License-Identifier: Apache-2.0
+# Copyright 2022 Stéphane Caron
+
+"""YAM URDF description."""
+
+from os import getenv as _getenv
+from os import path as _path
+
+from ._cache import clone_to_cache as _clone_to_cache
+
+REPOSITORY_PATH: str = _clone_to_cache(
+    "i2rt",
+    commit=_getenv("ROBOT_DESCRIPTION_COMMIT", None),
+)
+
+PACKAGE_PATH: str = _path.join(REPOSITORY_PATH, "robot_models/yam/assets")
+
+URDF_PATH: str = _path.join(REPOSITORY_PATH, "robot_models/yam/yam.urdf")