+++ robot_descriptions/_repositories.py
@@ -205,7 +205,7 @@ REPOSITORIES: Dict[str, Repository] = {
     ),
     "mujoco_menagerie": Repository(
         url="https://github.com/deepmind/mujoco_menagerie.git",
-        commit="66384c6b8581c811a7b1eb63bcf4fa944fa43602",
+        commit="c1503a62496b64222c64ff65dd652d461a5b064e",
         cache_path="mujoco_menagerie",
     ),
     "nao_robot": Repository(