+++ CHANGELOG.md
@@ -32,6 +32,8 @@ All notable changes to this project will be documented in this file.
 - Description: Fourier N1 (URDF)
 - Description: RSK omnidirectional robot (MJCF)
 - Description: YAM (MJCF) (thanks to @kevinzakka)
+- Description YAM (URDF)
+- Description RBY1 Mobile Manipulatior (URDF)
 
 ### Changed