+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
-- Description YAM (URDF)
-- Description RBY1 Mobile Manipulatior (URDF)
+
+- Description RBY1 Mobile Manipulatior (URDF) (thanks to @uynitsuj)
+- Description YAM (MJCF) (thanks to @uynitsuj)
+- Description YAM (URDF) (thanks to @uynitsuj)
 
 ## [1.19.0] - 2025-07-03