+++ CHANGELOG.md
@@ -6,8 +6,9 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- Description: Unitree H1\_2 (URDF) (thanks to @TonyZYT2000)
+- Description: ToddlerBot (URDF)
 - Description: Unitree H1\_2 (MJCF) (thanks to @TonyZYT2000)
+- Description: Unitree H1\_2 (URDF) (thanks to @TonyZYT2000)
 
 ### Changed